requirements.txt
streamlit
sentence-transformers
scikit-learn
email-validator
PyPDF2
python-docx
streamlit-authenticator
transformers==4.33.2
torch==2.0.1
tensorflow==2.13.0
 